Update and share libc++ between ports along with misc fixes
* Imagine: Add support for and use a bundled libc++ on Android and iOS
* Imagine: Fix build errors on libc++ 14.0
* Imagine: Make use of C++20 <concepts> library and C++23 string_view::contains(), to_underlying() from newer libc++
* Imagine: Replace iterateTimes* macros with iota from <ranges>
* Imagine: Clean up various Renderer functions and enums
* Imagine: Move -ffunction-sections and -fdata-sections compile flags into common makefile
* Imagine: Disable explicit visibility attributes when building liblzma to prevent unwanted exported symbols
* Imagine: Set animator to finished value in FrameAnimation::cancel()
* Imagine: Set buildPath in makefile when building bundled libraries
* EmuFramework: Ignore duplicate input device settings from config file
* EmuFramework: Don't export mpcdec symbols from Mednafen in build output
* EmuFramework: Remove unused variable EmuSystem::canRenderRGB565
* C64.emu: Fix RGB565 color rendering broken from last VICE core update
* Snes9x EX+: Work around name conflict from MacTypes.h
* Update CI workflow to use Android NDK r25

+++ imagine/include/imagine/gfx/PixmapTexture.hh
@@ -45,7 +45,7 @@ public:
 	using PixmapTextureImpl::PixmapTextureImpl;
 	PixmapTexture(RendererTask &, TextureConfig, IG::ErrorCode *errorPtr = nullptr);
 	PixmapTexture(RendererTask &, IG::Data::PixmapSource, const TextureSampler *compatSampler, bool makeMipmaps, IG::ErrorCode *errorPtr = nullptr);
-	IG::ErrorCode setFormat(IG::PixmapDesc desc, uint8_t levels, ColorSpace c = {}, const TextureSampler *compatSampler = {});
+	IG::ErrorCode setFormat(IG::PixmapDesc desc, int levels, ColorSpace c = {}, const TextureSampler *compatSampler = {});
 	FRect uvBounds() const;
 	IG::PixmapDesc usedPixmapDesc() const;
 	operator TextureSpan() const;
